Module aws_sdk_appflow::error

Modules

conflict_exception

See ConflictException

connector_authentication_exception

See ConnectorAuthenticationException

connector_server_exception

See ConnectorServerException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

unsupported_operation_exception

See UnsupportedOperationException

validation_exception

See ValidationException

Structs

ConflictException

There was a conflict when processing the request (for example, a flow with the given name

ConnectorAuthenticationException

An error occurred when authenticating with the connector endpoint.

ConnectorServerException

An error occurred when retrieving data from the connector endpoint.

CreateConnectorProfileError

CreateFlowError

DeleteConnectorProfileError

DeleteFlowError

DescribeConnectorEntityError

DescribeConnectorProfilesError

DescribeConnectorsError

DescribeFlowError

DescribeFlowExecutionRecordsError

InternalServerException

An internal service error occurred during the processing of your request. Try again

ListConnectorEntitiesError

ListFlowsError

ListTagsForResourceError

ResourceNotFoundException

The resource specified in the request (such as the source or destination connector

ServiceQuotaExceededException

The request would cause a service quota (such as the number of flows) to be exceeded.

StartFlowError

StopFlowError

TagResourceError

UnsupportedOperationException

The requested operation is not supported for the current flow.

UntagResourceError

UpdateConnectorProfileError

UpdateFlowError

ValidationException

The request has invalid or missing parameters.

Enums

CreateConnectorProfileErrorKind

CreateFlowErrorKind

DeleteConnectorProfileErrorKind

DeleteFlowErrorKind

DescribeConnectorEntityErrorKind

DescribeConnectorProfilesErrorKind

DescribeConnectorsErrorKind

DescribeFlowErrorKind

DescribeFlowExecutionRecordsErrorKind

ListConnectorEntitiesErrorKind

ListFlowsErrorKind

ListTagsForResourceErrorKind

StartFlowErrorKind

StopFlowErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateConnectorProfileErrorKind

UpdateFlowErrorKind