Module aws_sdk_appflow::model

Modules

aggregation_config

See AggregationConfig

amplitude_connector_profile_credentials

See AmplitudeConnectorProfileCredentials

amplitude_connector_profile_properties

See AmplitudeConnectorProfileProperties

amplitude_metadata

See AmplitudeMetadata

amplitude_source_properties

See AmplitudeSourceProperties

basic_auth_credentials

See BasicAuthCredentials

connector_configuration

See ConnectorConfiguration

connector_entity

See ConnectorEntity

connector_entity_field

See ConnectorEntityField

connector_metadata

See ConnectorMetadata

connector_o_auth_request

See ConnectorOAuthRequest

connector_operator

See ConnectorOperator

connector_profile

See ConnectorProfile

connector_profile_config

See ConnectorProfileConfig

connector_profile_credentials

See ConnectorProfileCredentials

connector_profile_properties

See ConnectorProfileProperties

customer_profiles_destination_properties

See CustomerProfilesDestinationProperties

customer_profiles_metadata

See CustomerProfilesMetadata

datadog_connector_profile_credentials

See DatadogConnectorProfileCredentials

datadog_connector_profile_properties

See DatadogConnectorProfileProperties

datadog_metadata

See DatadogMetadata

datadog_source_properties

See DatadogSourceProperties

destination_connector_properties

See DestinationConnectorProperties

destination_field_properties

See DestinationFieldProperties

destination_flow_config

See DestinationFlowConfig

dynatrace_connector_profile_credentials

See DynatraceConnectorProfileCredentials

dynatrace_connector_profile_properties

See DynatraceConnectorProfileProperties

dynatrace_metadata

See DynatraceMetadata

dynatrace_source_properties

See DynatraceSourceProperties

error_handling_config

See ErrorHandlingConfig

error_info

See ErrorInfo

event_bridge_destination_properties

See EventBridgeDestinationProperties

event_bridge_metadata

See EventBridgeMetadata

execution_details

See ExecutionDetails

execution_record

See ExecutionRecord

execution_result

See ExecutionResult

field_type_details

See FieldTypeDetails

flow_definition

See FlowDefinition

google_analytics_connector_profile_credentials

See GoogleAnalyticsConnectorProfileCredentials

google_analytics_connector_profile_properties

See GoogleAnalyticsConnectorProfileProperties

google_analytics_metadata

See GoogleAnalyticsMetadata

google_analytics_source_properties

See GoogleAnalyticsSourceProperties

honeycode_connector_profile_credentials

See HoneycodeConnectorProfileCredentials

honeycode_connector_profile_properties

See HoneycodeConnectorProfileProperties

honeycode_destination_properties

See HoneycodeDestinationProperties

honeycode_metadata

See HoneycodeMetadata

incremental_pull_config

See IncrementalPullConfig

infor_nexus_connector_profile_credentials

See InforNexusConnectorProfileCredentials

infor_nexus_connector_profile_properties

See InforNexusConnectorProfileProperties

infor_nexus_metadata

See InforNexusMetadata

infor_nexus_source_properties

See InforNexusSourceProperties

lookout_metrics_destination_properties

See LookoutMetricsDestinationProperties

marketo_connector_profile_credentials

See MarketoConnectorProfileCredentials

marketo_connector_profile_properties

See MarketoConnectorProfileProperties

marketo_metadata

See MarketoMetadata

marketo_source_properties

See MarketoSourceProperties

o_auth_credentials

See OAuthCredentials

o_auth_properties

See OAuthProperties

prefix_config

See PrefixConfig

private_connection_provisioning_state

See PrivateConnectionProvisioningState

redshift_connector_profile_credentials

See RedshiftConnectorProfileCredentials

redshift_connector_profile_properties

See RedshiftConnectorProfileProperties

redshift_destination_properties

See RedshiftDestinationProperties

redshift_metadata

See RedshiftMetadata

s3_destination_properties

See S3DestinationProperties

s3_metadata

See S3Metadata

s3_output_format_config

See S3OutputFormatConfig

s3_source_properties

See S3SourceProperties

salesforce_connector_profile_credentials

See SalesforceConnectorProfileCredentials

salesforce_connector_profile_properties

See SalesforceConnectorProfileProperties

salesforce_destination_properties

See SalesforceDestinationProperties

salesforce_metadata

See SalesforceMetadata

salesforce_source_properties

See SalesforceSourceProperties

sapo_data_connector_profile_credentials

See SapoDataConnectorProfileCredentials

sapo_data_connector_profile_properties

See SapoDataConnectorProfileProperties

sapo_data_metadata

See SapoDataMetadata

sapo_data_source_properties

See SapoDataSourceProperties

scheduled_trigger_properties

See ScheduledTriggerProperties

service_now_connector_profile_credentials

See ServiceNowConnectorProfileCredentials

service_now_connector_profile_properties

See ServiceNowConnectorProfileProperties

service_now_metadata

See ServiceNowMetadata

service_now_source_properties

See ServiceNowSourceProperties

singular_connector_profile_credentials

See SingularConnectorProfileCredentials

singular_connector_profile_properties

See SingularConnectorProfileProperties

singular_metadata

See SingularMetadata

singular_source_properties

See SingularSourceProperties

slack_connector_profile_credentials

See SlackConnectorProfileCredentials

slack_connector_profile_properties

See SlackConnectorProfileProperties

slack_metadata

See SlackMetadata

slack_source_properties

See SlackSourceProperties

snowflake_connector_profile_credentials

See SnowflakeConnectorProfileCredentials

snowflake_connector_profile_properties

See SnowflakeConnectorProfileProperties

snowflake_destination_properties

See SnowflakeDestinationProperties

snowflake_metadata

See SnowflakeMetadata

source_connector_properties

See SourceConnectorProperties

source_field_properties

See SourceFieldProperties

source_flow_config

See SourceFlowConfig

supported_field_type_details

See SupportedFieldTypeDetails

task

See Task

trendmicro_connector_profile_credentials

See TrendmicroConnectorProfileCredentials

trendmicro_connector_profile_properties

See TrendmicroConnectorProfileProperties

trendmicro_metadata

See TrendmicroMetadata

trendmicro_source_properties

See TrendmicroSourceProperties

trigger_config

See TriggerConfig

trigger_properties

See TriggerProperties

upsolver_destination_properties

See UpsolverDestinationProperties

upsolver_metadata

See UpsolverMetadata

upsolver_s3_output_format_config

See UpsolverS3OutputFormatConfig

veeva_connector_profile_credentials

See VeevaConnectorProfileCredentials

veeva_connector_profile_properties

See VeevaConnectorProfileProperties

veeva_metadata

See VeevaMetadata

veeva_source_properties

See VeevaSourceProperties

zendesk_connector_profile_credentials

See ZendeskConnectorProfileCredentials

zendesk_connector_profile_properties

See ZendeskConnectorProfileProperties

zendesk_destination_properties

See ZendeskDestinationProperties

zendesk_metadata

See ZendeskMetadata

zendesk_source_properties

See ZendeskSourceProperties

Structs

AggregationConfig

The aggregation settings that you can use to customize the output format of your flow

AmplitudeConnectorProfileCredentials

The connector-specific credentials required when using Amplitude.

AmplitudeConnectorProfileProperties

The connector-specific profile properties required when using Amplitude.

AmplitudeMetadata

The connector metadata specific to Amplitude.

AmplitudeSourceProperties

The properties that are applied when Amplitude is being used as a source.

BasicAuthCredentials

The basic auth credentials required for basic authentication.

ConnectorConfiguration

The configuration settings related to a given connector.

ConnectorEntity

The high-level entity that can be queried in Amazon AppFlow. For example, a Salesforce

ConnectorEntityField

Describes the data model of a connector field. For example, for an

ConnectorMetadata

A structure to specify connector-specific metadata such as oAuthScopes,

ConnectorOAuthRequest

Used by select connectors for which the OAuth workflow is supported, such as Salesforce,

ConnectorOperator

The operation to be performed on the provided source fields.

ConnectorProfile

Describes an instance of a connector. This includes the provided name, credentials ARN,

ConnectorProfileConfig

Defines the connector-specific configuration and credentials for the connector profile.

ConnectorProfileCredentials

The connector-specific credentials required by a connector.

ConnectorProfileProperties

The connector-specific profile properties required by each connector.

CustomerProfilesDestinationProperties

The properties that are applied when Amazon Connect Customer Profiles is used as a

CustomerProfilesMetadata

The connector metadata specific to Amazon Connect Customer Profiles.

DatadogConnectorProfileCredentials

The connector-specific credentials required by Datadog.

DatadogConnectorProfileProperties

The connector-specific profile properties required by Datadog.

DatadogMetadata

The connector metadata specific to Datadog.

DatadogSourceProperties

The properties that are applied when Datadog is being used as a source.

DestinationConnectorProperties

This stores the information that is required to query a particular connector.

DestinationFieldProperties

The properties that can be applied to a field when connector is being used as a

DestinationFlowConfig

Contains information about the configuration of destination connectors present in the

DynatraceConnectorProfileCredentials

The connector-specific profile credentials required by Dynatrace.

DynatraceConnectorProfileProperties

The connector-specific profile properties required by Dynatrace.

DynatraceMetadata

The connector metadata specific to Dynatrace.

DynatraceSourceProperties

The properties that are applied when Dynatrace is being used as a source.

ErrorHandlingConfig

The settings that determine how Amazon AppFlow handles an error when placing data in the

ErrorInfo

Provides details in the event of a failed flow, including the failure count and the

EventBridgeDestinationProperties

The properties that are applied when Amazon EventBridge is being used as a destination.

EventBridgeMetadata

The connector metadata specific to Amazon EventBridge.

ExecutionDetails

Describes the details of the flow run, including the timestamp, status, and message.

ExecutionRecord

Specifies information about the past flow run instances for a given flow.

ExecutionResult

Specifies the end result of the flow run.

FieldTypeDetails

Contains details regarding the supported field type and the operators that can be applied

FlowDefinition

The properties of the flow, such as its source, destination, trigger type, and so on.

GoogleAnalyticsConnectorProfileCredentials

The connector-specific profile credentials required by Google Analytics.

GoogleAnalyticsConnectorProfileProperties

The connector-specific profile properties required by Google Analytics.

GoogleAnalyticsMetadata

The connector metadata specific to Google Analytics.

GoogleAnalyticsSourceProperties

The properties that are applied when Google Analytics is being used as a source.

HoneycodeConnectorProfileCredentials

The connector-specific credentials required when using Amazon Honeycode.

HoneycodeConnectorProfileProperties

The connector-specific properties required when using Amazon Honeycode.

HoneycodeDestinationProperties

The properties that are applied when Amazon Honeycode is used as a destination.

HoneycodeMetadata

The connector metadata specific to Amazon Honeycode.

IncrementalPullConfig

Specifies the configuration used when importing incremental records from the source.

InforNexusConnectorProfileCredentials

The connector-specific profile credentials required by Infor Nexus.

InforNexusConnectorProfileProperties

The connector-specific profile properties required by Infor Nexus.

InforNexusMetadata

The connector metadata specific to Infor Nexus.

InforNexusSourceProperties

The properties that are applied when Infor Nexus is being used as a source.

LookoutMetricsDestinationProperties

The properties that are applied when Amazon Lookout for Metrics is used as a destination.

MarketoConnectorProfileCredentials

The connector-specific profile credentials required by Marketo.

MarketoConnectorProfileProperties

The connector-specific profile properties required when using Marketo.

MarketoMetadata

The connector metadata specific to Marketo.

MarketoSourceProperties

The properties that are applied when Marketo is being used as a source.

OAuthCredentials

The OAuth credentials required for OAuth type authentication.

OAuthProperties

The OAuth properties required for OAuth type authentication.

PrefixConfig

Determines the prefix that Amazon AppFlow applies to the destination folder name. You can

PrivateConnectionProvisioningState

Specifies the private connection provisioning state.

RedshiftConnectorProfileCredentials

The connector-specific profile credentials required when using Amazon Redshift.

RedshiftConnectorProfileProperties

The connector-specific profile properties when using Amazon Redshift.

RedshiftDestinationProperties

The properties that are applied when Amazon Redshift is being used as a destination.

RedshiftMetadata

The connector metadata specific to Amazon Redshift.

S3DestinationProperties

The properties that are applied when Amazon S3 is used as a destination.

S3Metadata

The connector metadata specific to Amazon S3.

S3OutputFormatConfig

The configuration that determines how Amazon AppFlow should format the flow output data

S3SourceProperties

The properties that are applied when Amazon S3 is being used as the flow source.

SalesforceConnectorProfileCredentials

The connector-specific profile credentials required when using Salesforce.

SalesforceConnectorProfileProperties

The connector-specific profile properties required when using Salesforce.

SalesforceDestinationProperties

The properties that are applied when Salesforce is being used as a destination.

SalesforceMetadata

The connector metadata specific to Salesforce.

SalesforceSourceProperties

The properties that are applied when Salesforce is being used as a source.

SapoDataConnectorProfileCredentials

The connector-specific profile credentials required when using SAPOData.

SapoDataConnectorProfileProperties

The connector-specific profile properties required when using SAPOData.

SapoDataMetadata

The connector metadata specific to SAPOData.

SapoDataSourceProperties

The properties that are applied when using SAPOData as a flow source.

ScheduledTriggerProperties

Specifies the configuration details of a schedule-triggered flow as defined by the user.

ServiceNowConnectorProfileCredentials

The connector-specific profile credentials required when using ServiceNow.

ServiceNowConnectorProfileProperties

The connector-specific profile properties required when using ServiceNow.

ServiceNowMetadata

The connector metadata specific to ServiceNow.

ServiceNowSourceProperties

The properties that are applied when ServiceNow is being used as a source.

SingularConnectorProfileCredentials

The connector-specific profile credentials required when using Singular.

SingularConnectorProfileProperties

The connector-specific profile properties required when using Singular.

SingularMetadata

The connector metadata specific to Singular.

SingularSourceProperties

The properties that are applied when Singular is being used as a source.

SlackConnectorProfileCredentials

The connector-specific profile credentials required when using Slack.

SlackConnectorProfileProperties

The connector-specific profile properties required when using Slack.

SlackMetadata

The connector metadata specific to Slack.

SlackSourceProperties

The properties that are applied when Slack is being used as a source.

SnowflakeConnectorProfileCredentials

The connector-specific profile credentials required when using Snowflake.

SnowflakeConnectorProfileProperties

The connector-specific profile properties required when using Snowflake.

SnowflakeDestinationProperties

The properties that are applied when Snowflake is being used as a destination.

SnowflakeMetadata

The connector metadata specific to Snowflake.

SourceConnectorProperties

Specifies the information that is required to query a particular connector.

SourceFieldProperties

The properties that can be applied to a field when the connector is being used as a

SourceFlowConfig

Contains information about the configuration of the source connector used in the flow.

SupportedFieldTypeDetails

Contains details regarding all the supported FieldTypes and their

Task

A class for modeling different type of tasks. Task implementation varies based on the

TrendmicroConnectorProfileCredentials

The connector-specific profile credentials required when using Trend Micro.

TrendmicroConnectorProfileProperties

The connector-specific profile properties required when using Trend Micro.

TrendmicroMetadata

The connector metadata specific to Trend Micro.

TrendmicroSourceProperties

The properties that are applied when using Trend Micro as a flow source.

TriggerConfig

The trigger settings that determine how and when Amazon AppFlow runs the specified flow.

TriggerProperties

Specifies the configuration details that control the trigger for a flow. Currently, these

UpsolverDestinationProperties

The properties that are applied when Upsolver is used as a destination.

UpsolverMetadata

The connector metadata specific to Upsolver.

UpsolverS3OutputFormatConfig

The configuration that determines how Amazon AppFlow formats the flow output data when

VeevaConnectorProfileCredentials

The connector-specific profile credentials required when using Veeva.

VeevaConnectorProfileProperties

The connector-specific profile properties required when using Veeva.

VeevaMetadata

The connector metadata specific to Veeva.

VeevaSourceProperties

The properties that are applied when using Veeva as a flow source.

ZendeskConnectorProfileCredentials

The connector-specific profile credentials required when using Zendesk.

ZendeskConnectorProfileProperties

The connector-specific profile properties required when using Zendesk.

ZendeskDestinationProperties

The properties that are applied when Zendesk is used as a destination.

ZendeskMetadata

The connector metadata specific to Zendesk.

ZendeskSourceProperties

The properties that are applied when using Zendesk as a flow source.

Enums

AggregationType

AmplitudeConnectorOperator

ConnectionMode

ConnectorType

DataPullMode

DatadogConnectorOperator

DynatraceConnectorOperator

ExecutionStatus

FileType

FlowStatus

GoogleAnalyticsConnectorOperator

InforNexusConnectorOperator

MarketoConnectorOperator

Operator

OperatorPropertiesKeys

PrefixFormat

PrefixType

PrivateConnectionProvisioningFailureCause

PrivateConnectionProvisioningStatus

S3ConnectorOperator

SalesforceConnectorOperator

SapoDataConnectorOperator

ScheduleFrequencyType

ServiceNowConnectorOperator

SingularConnectorOperator

SlackConnectorOperator

TaskType

TrendmicroConnectorOperator

TriggerType

VeevaConnectorOperator

WriteOperationType

The possible write operations in the destination connector. When this value is not

ZendeskConnectorOperator