Struct aws_sdk_appflow::model::connector_operator::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ConnectorOperator

Implementations

impl Builder

pub fn amplitude(self, input: AmplitudeConnectorOperator) -> Self

The operation to be performed on the provided Amplitude source fields.

pub fn set_amplitude(self, input: Option<AmplitudeConnectorOperator>) -> Self

pub fn datadog(self, input: DatadogConnectorOperator) -> Self

The operation to be performed on the provided Datadog source fields.

pub fn set_datadog(self, input: Option<DatadogConnectorOperator>) -> Self

pub fn dynatrace(self, input: DynatraceConnectorOperator) -> Self

The operation to be performed on the provided Dynatrace source fields.

pub fn set_dynatrace(self, input: Option<DynatraceConnectorOperator>) -> Self

pub fn google_analytics(self, input: GoogleAnalyticsConnectorOperator) -> Self

The operation to be performed on the provided Google Analytics source fields.

pub fn set_google_analytics(
self,
input: Option<GoogleAnalyticsConnectorOperator>
) -> Self

pub fn infor_nexus(self, input: InforNexusConnectorOperator) -> Self

The operation to be performed on the provided Infor Nexus source fields.

pub fn set_infor_nexus(self, input: Option<InforNexusConnectorOperator>) -> Self

pub fn marketo(self, input: MarketoConnectorOperator) -> Self

The operation to be performed on the provided Marketo source fields.

pub fn set_marketo(self, input: Option<MarketoConnectorOperator>) -> Self

pub fn s3(self, input: S3ConnectorOperator) -> Self

The operation to be performed on the provided Amazon S3 source fields.

pub fn set_s3(self, input: Option<S3ConnectorOperator>) -> Self

pub fn salesforce(self, input: SalesforceConnectorOperator) -> Self

The operation to be performed on the provided Salesforce source fields.

pub fn set_salesforce(self, input: Option<SalesforceConnectorOperator>) -> Self

pub fn service_now(self, input: ServiceNowConnectorOperator) -> Self

The operation to be performed on the provided ServiceNow source fields.

pub fn set_service_now(self, input: Option<ServiceNowConnectorOperator>) -> Self

pub fn singular(self, input: SingularConnectorOperator) -> Self

The operation to be performed on the provided Singular source fields.

pub fn set_singular(self, input: Option<SingularConnectorOperator>) -> Self

pub fn slack(self, input: SlackConnectorOperator) -> Self

The operation to be performed on the provided Slack source fields.

pub fn set_slack(self, input: Option<SlackConnectorOperator>) -> Self

pub fn trendmicro(self, input: TrendmicroConnectorOperator) -> Self

The operation to be performed on the provided Trend Micro source fields.

pub fn set_trendmicro(self, input: Option<TrendmicroConnectorOperator>) -> Self

pub fn veeva(self, input: VeevaConnectorOperator) -> Self

The operation to be performed on the provided Veeva source fields.

pub fn set_veeva(self, input: Option<VeevaConnectorOperator>) -> Self

pub fn zendesk(self, input: ZendeskConnectorOperator) -> Self

The operation to be performed on the provided Zendesk source fields.

pub fn set_zendesk(self, input: Option<ZendeskConnectorOperator>) -> Self

pub fn sapo_data(self, input: SapoDataConnectorOperator) -> Self

The operation to be performed on the provided SAPOData source fields.

pub fn set_sapo_data(self, input: Option<SapoDataConnectorOperator>) -> Self

pub fn build(self) -> ConnectorOperator

Consumes the builder and constructs a ConnectorOperator