Module aws_sdk_appflow::operation

Structs

CreateConnectorProfile

Creates a new connector profile associated with your Amazon Web Services account. There is a soft quota

CreateFlow

Enables your application to create a new flow using Amazon AppFlow. You must create a

DeleteConnectorProfile

Enables you to delete an existing connector profile.

DeleteFlow

Enables your application to delete an existing flow. Before deleting the flow, Amazon

DescribeConnectorEntity

Provides details regarding the entity used with the connector, with a description of the

DescribeConnectorProfiles

Returns a list of connector-profile details matching the provided

DescribeConnectors

Describes the connectors vended by Amazon AppFlow for specified connector types. If you

DescribeFlow

Provides a description of the specified flow.

DescribeFlowExecutionRecords

Fetches the execution history of the flow.

ListConnectorEntities

Returns the list of available connector entities supported by Amazon AppFlow. For

ListFlows

Lists all of the flows associated with your account.

ListTagsForResource

Retrieves the tags that are associated with a specified flow.

StartFlow

Activates an existing flow. For on-demand flows, this operation runs the flow

StopFlow

Deactivates the existing flow. For on-demand flows, this operation returns an

TagResource

Applies a tag to the specified flow.

UntagResource

Removes a tag from the specified flow.

UpdateConnectorProfile

Updates a given connector profile associated with your account.

UpdateFlow

Updates an existing flow.